+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#3.0.1
+- Bug fix for custom search buttons
+- Changed default URL for Wikipedia search
+- Bug fix for 'move up' reveal animation - now tooltip is not interactive in first half of animation duration
+
 #3.0.0
 
 New: